Fixing code for database, add code for self assessment feature

+++ database/migrations/2025_05_25_141054_create_reviews_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->increments('id_review'); // Primary key auto increment
             $table->unsignedInteger('id_user');      // FK ke users
             $table->unsignedInteger('id_dokter');    // FK ke users (dokter)
-            $table->unsignedBigInteger('id_konsul'); // FK ke consultations (asumsi tipe bigint)
+            $table->unsignedInteger('id_konsul'); // FK ke consultations (asumsi tipe bigint)
 
             $table->date('tanggal_review');
             $table->time('waktu_review');
